requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv
requests
asyncpg
asyncio
pillow
firebase_admin